+++ src/draw.c
@@ -42,7 +42,7 @@
  * \param b Blue
  */
 void draw_line(unsigned char img[],
-               unsigned int width, unsigned int height,
+               int width, int height,
                int bitsperpixel,
                int tx, int ty, int bx, int by,
                int line_width,
@@ -114,7 +114,7 @@ void draw_line(unsigned char img[],
  * \param b Blue
  */
 void draw_dot(unsigned char img[],
-              unsigned int width, unsigned int height,
+              int width, int height,
               int bitsperpixel,
               int centre_x, int centre_y, int radius,
               int r, int g, int b)
@@ -127,16 +127,16 @@ void draw_dot(unsigned char img[],
     int by = centre_y + radius;
 
     if (tx < 0) tx = 0;
-    if (bx >= (int)width) bx = (int)width-1;
+    if (bx >= width) bx = width-1;
     if (ty < 0) ty = 0;
-    if (by >= (int)height) by = (int)height-1;
+    if (by >= height) by = height-1;
 
     for (y = ty; y <= by; y++) {
         dy = y - centre_y;
         for (x = tx; x <= bx; x++) {
             dx = x - centre_x;
             if (SQUARE_MAG(dx, dy) > radius_sqr) continue;
-            n = (y * (int)width + x) * bytesperpixel;
+            n = (y * width + x) * bytesperpixel;
             img[n+2] = (unsigned char)b;
             img[n+1] = (unsigned char)g;
             img[n] = (unsigned char)r;
